# src/ShulamitP_Mizrahi_Calc_Api/Models/OperationType.cs
/*
 * Mizrahi Arithmetic API
 *
 * EST API that performs basic arithmetic operations. Implemented by Shulamit Petranker for Matrix/BM assignment.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace ShulamitP_Mizrahi_Calc_Api.Models
{
  
          /// <summary>
          /// Gets or Sets OperationType
          /// </summary>
          [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
          public enum OperationType
          {
              /// <summary>
              /// Enum AddEnum for Add
              /// </summary>
              [EnumMember(Value = "Add")]
              Add = 0,
              /// <summary>
              /// Enum SubtractEnum for Subtract
              /// </summary>
              [EnumMember(Value = "Subtract")]
              Subtract = 1,
              /// <summary>
              /// Enum MultiplyEnum for Multiply
              /// </summary>
              [EnumMember(Value = "Multiply")]
              Multiply = 2,
              /// <summary>
              /// Enum DivideEnum for Divide
              /// </summary>
              [EnumMember(Value = "Divide")]
              Divide = 3          }
}